Convert Study to Hibernate and add StudyType model
JIRA: GLT-1283
