+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/Library.java
@@ -349,4 +349,5 @@ public interface Library extends SecurableByProfile, Comparable, Barcodable, Loc
   public IndexFamily getCurrentFamily();
 
   void setLastModified(Date lastModified);
+
 }
