+++ core/src/main/java/uk/ac/bbsrc/tgac/miso/core/data/impl/StudyImpl.java
@@ -24,14 +24,34 @@
 package uk.ac.bbsrc.tgac.miso.core.data.impl;
 
 import java.io.Serializable;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.HashSet;
 
+import javax.persistence.CascadeType;
+import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
+import javax.persistence.OneToOne;
+import javax.persistence.Table;
+import javax.persistence.Transient;
+
+import org.w3c.dom.Document;
 
 import com.eaglegenomics.simlims.core.SecurityProfile;
 import com.eaglegenomics.simlims.core.User;
 
-import uk.ac.bbsrc.tgac.miso.core.data.AbstractStudy;
+import uk.ac.bbsrc.tgac.miso.core.data.AbstractExperiment;
+import uk.ac.bbsrc.tgac.miso.core.data.ChangeLog;
+import uk.ac.bbsrc.tgac.miso.core.data.Experiment;
 import uk.ac.bbsrc.tgac.miso.core.data.Project;
+import uk.ac.bbsrc.tgac.miso.core.data.Study;
+import uk.ac.bbsrc.tgac.miso.core.exception.MalformedExperimentException;
+import uk.ac.bbsrc.tgac.miso.core.security.SecurableByProfile;
 
 /**
  * Concrete implementation of a Study
@@ -40,23 +60,48 @@ import uk.ac.bbsrc.tgac.miso.core.data.Project;
  * @since 0.0.2
  */
 @Entity
-public class StudyImpl extends AbstractStudy implements Serializable {
+@Table(name = "`Study`")
+public class StudyImpl implements Study, Serializable {
+  public static final Long UNSAVED_ID = 0L;
+
+  private static final long serialVersionUID = 1L;
+
+  @ManyToOne(cascade = CascadeType.ALL)
+  private Project project = null;
+
+  @OneToMany(targetEntity = AbstractExperiment.class, cascade = CascadeType.ALL)
+  private Collection<Experiment> experiments = new HashSet<>();
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.AUTO)
+  private long studyId = StudyImpl.UNSAVED_ID;
+
+  @Transient
+  public Document submissionDocument;
+
+  @OneToOne(cascade = CascadeType.ALL)
+  private SecurityProfile securityProfile = null;
+
+  @Column(name = "name", nullable = false)
+  private String name;
+  @Column(name = "description", nullable = false)
+  private String description;
+  @Column(name = "accession")
+  private String accession;
+  @Column(name = "abstract")
+  private String abs;
+  @Column(name = "studyType")
+  private String studyType;
+  @Column(name = "alias")
+  private String alias;
+  private final Collection<ChangeLog> changeLog = new ArrayList<>();
+  private User lastModifier;
 
   /**
    * Construct a new Study with a default empty SecurityProfile
    */
   public StudyImpl() {
     setSecurityProfile(new SecurityProfile());
-  }
-
-  /**
-   * Construct a new Study with a SecurityProfile owned by the given User
-   * 
-   * @param user
-   *          of type User
-   */
-  public StudyImpl(User user) {
-    setSecurityProfile(new SecurityProfile(user));
   }
 
   /**
@@ -77,7 +122,227 @@ public class StudyImpl extends AbstractStudy implements Serializable {
     }
   }
 
+  /**
+   * Construct a new Study with a SecurityProfile owned by the given User
+   * 
+   * @param user
+   *          of type User
+   */
+  public StudyImpl(User user) {
+    setSecurityProfile(new SecurityProfile(user));
+  }
+
+  @Override
+  public void addExperiment(Experiment e) throws MalformedExperimentException {
+    // propagate security profiles down the hierarchy
+    e.setSecurityProfile(this.securityProfile);
+
+    // add
+    this.experiments.add(e);
+  }
+
   @Override
   public void buildSubmission() {
+  }
+
+  @Override
+  public int compareTo(Object o) {
+    Study t = (Study) o;
+    if (getId() < t.getId()) return -1;
+    if (getId() > t.getId()) return 1;
+    return 0;
+  }
+
+  /**
+   * Equivalency is based on getProjectId() if set, otherwise on name, description and creation date.
+   */
+  @Override
+  public boolean equals(Object obj) {
+    if (obj == null) return false;
+    if (obj == this) return true;
+    if (!(obj instanceof Study)) return false;
+    Study them = (Study) obj;
+    // If not saved, then compare resolved actual objects. Otherwise
+    // just compare IDs.
+    if (getId() == StudyImpl.UNSAVED_ID || them.getId() == StudyImpl.UNSAVED_ID) {
+      if (getName() != null && them.getName() != null) {
+        return getName().equals(them.getName());
+      } else {
+        return getAlias().equals(them.getAlias());
+      }
+    } else {
+      return this.getId() == them.getId();
+    }
+  }
+
+  @Override
+  public String getAbstract() {
+    return abs;
+  }
+
+  @Override
+  public String getAccession() {
+    return accession;
+  }
+
+  @Override
+  public String getAlias() {
+    return alias;
+  }
+
+  @Override
+  public Collection<ChangeLog> getChangeLog() {
+    return changeLog;
+  }
+
+  @Override
+  public String getDescription() {
+    return description;
+  }
+
+  @Override
+  public Collection<Experiment> getExperiments() {
+    return experiments;
+  }
+
+  @Override
+  public long getId() {
+    return studyId;
+  }
+
+  @Override
+  public User getLastModifier() {
+    return lastModifier;
+  }
+
+  @Override
+  public String getName() {
+    return name;
+  }
+
+  @Override
+  public Project getProject() {
+    return project;
+  }
+
+  @Override
+  public SecurityProfile getSecurityProfile() {
+    return securityProfile;
+  }
+
+  @Override
+  public String getStudyType() {
+    return studyType;
+  }
+
+  @Override
+  public int hashCode() {
+    if (this.getId() != StudyImpl.UNSAVED_ID) {
+      return (int) getId();
+    } else {
+      final int PRIME = 37;
+      int hashcode = 1;
+      if (getName() != null) hashcode = PRIME * hashcode + getName().hashCode();
+      if (getAlias() != null) hashcode = PRIME * hashcode + getAlias().hashCode();
+      return hashcode;
+    }
+  }
+
+  @Override
+  public void inheritPermissions(SecurableByProfile parent) throws SecurityException {
+    if (parent.getSecurityProfile().getOwner() != null) {
+      setSecurityProfile(parent.getSecurityProfile());
+    } else {
+      throw new SecurityException("Cannot inherit permissions when parent object owner is not set!");
+    }
+  }
+
+  @Override
+  public boolean isDeletable() {
+    return getId() != StudyImpl.UNSAVED_ID && getExperiments().isEmpty();
+  }
+
+  @Override
+  public void setAbstract(String abs) {
+    this.abs = abs;
+  }
+
+  @Override
+  public void setAccession(String accession) {
+    this.accession = accession;
+  }
+
+  @Override
+  public void setAlias(String alias) {
+    this.alias = alias;
+  }
+
+  @Override
+  public void setDescription(String description) {
+    this.description = description;
+  }
+
+  @Override
+  public void setExperiments(Collection<Experiment> experiments) {
+    this.experiments = experiments;
+  }
+
+  @Override
+  public void setId(long id) {
+    this.studyId = id;
+  }
+
+  @Override
+  public void setLastModifier(User lastModifier) {
+    this.lastModifier = lastModifier;
+  }
+
+  @Override
+  public void setName(String name) {
+    this.name = name;
+  }
+
+  @Override
+  public void setProject(Project project) {
+    this.project = project;
+  }
+
+  @Override
+  public void setSecurityProfile(SecurityProfile securityProfile) {
+    this.securityProfile = securityProfile;
+  }
+
+  @Override
+  public void setStudyType(String studyType) {
+    this.studyType = studyType;
+  }
+
+  @Override
+  public String toString() {
+    StringBuilder sb = new StringBuilder();
+    sb.append(getName());
+    sb.append(" : ");
+    sb.append(getAlias());
+    sb.append(" : ");
+    sb.append(getDescription());
+    sb.append(" : ");
+    sb.append(getStudyType());
+    sb.append(" : ");
+
+    if (getProject() != null) {
+      sb.append(getProject().getAlias());
+      sb.append("(" + getProject().getName() + ")");
+    }
+    return sb.toString();
+  }
+
+  @Override
+  public boolean userCanRead(User user) {
+    return securityProfile.userCanRead(user);
+  }
+
+  @Override
+  public boolean userCanWrite(User user) {
+    return securityProfile.userCanWrite(user);
   }
 }
