+++ miso-web/src/main/webapp/WEB-INF/db-config.xml
@@ -46,6 +46,7 @@
         <value>uk.ac.bbsrc.tgac.miso.core.data.LibraryDesign</value>
         <value>uk.ac.bbsrc.tgac.miso.core.data.LibraryDesignCode</value>
         <value>uk.ac.bbsrc.tgac.miso.core.data.PoolOrderCompletion</value>
+        <value>uk.ac.bbsrc.tgac.miso.core.data.StudyType</value>
         <value>uk.ac.bbsrc.tgac.miso.core.data.impl.DetailedQcStatusImpl</value>
         <value>uk.ac.bbsrc.tgac.miso.core.data.impl.DetailedSampleImpl</value>
         <value>uk.ac.bbsrc.tgac.miso.core.data.impl.IdentityImpl</value>
@@ -70,6 +71,7 @@
         <value>uk.ac.bbsrc.tgac.miso.core.data.impl.SampleTissueProcessingImpl</value>
         <value>uk.ac.bbsrc.tgac.miso.core.data.impl.SampleValidRelationshipImpl</value>
         <value>uk.ac.bbsrc.tgac.miso.core.data.impl.SequencingParametersImpl</value>
+        <value>uk.ac.bbsrc.tgac.miso.core.data.impl.StudyImpl</value>
         <value>uk.ac.bbsrc.tgac.miso.core.data.impl.SubprojectImpl</value>
         <value>uk.ac.bbsrc.tgac.miso.core.data.impl.TissueMaterialImpl</value>
         <value>uk.ac.bbsrc.tgac.miso.core.data.impl.TissueOriginImpl</value>
