+++ spring/src/main/java/uk/ac/bbsrc/tgac/miso/spring/ajax/RunControllerHelperService.java
@@ -12,11 +12,11 @@
  *
  * MISO is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with MISO.  If not, see <http://www.gnu.org/licenses/>.
+ * along with MISO. If not, see <http://www.gnu.org/licenses/>.
  *
  * *********************************************************************
  */
@@ -1018,14 +1018,7 @@ public class RunControllerHelperService {
         // check if each poolable has been in a study for this pool already
         Collection<Dilution> ds = p.getPoolableElements();
         for (Dilution d : ds) {
-          Collection<Study> studies = requestManager.listAllStudiesByLibraryId(d.getLibrary().getId());
-          if (studies.isEmpty()) {
-            pooledProjects.add(d.getLibrary().getSample().getProject());
-          } else {
-            for (Study stu : studies) {
-              pooledProjects.add(stu.getProject());
-            }
-          }
+          pooledProjects.add(d.getLibrary().getSample().getProject());
         }
 
         for (Experiment poolExp : p.getExperiments()) {
