+++ sqlstore/src/main/resources/db/migration/V8900__hibernatize_everything.sql
@@ -1,3 +1,9 @@
 UPDATE KitDescriptor SET kitType = UPPER(kitType), platformType = UPPER(platformType);
 
 UPDATE Platform SET name = UPPER(name);
+
+ALTER TABLE Study ADD COLUMN studyTypeId bigint(20);
+UPDATE Study SET studyTypeId = (SELECT typeId FROM StudyType WHERE name = studyType);
+ALTER TABLE Study ADD CONSTRAINT study_studyTypeId FOREIGN KEY (studyTypeId) REFERENCES StudyType(typeId);
+ALTER TABLE Study DROP COLUMN studyType;
+ALTER TABLE Study ALTER COLUMN studyTypeId bigint(20) NOT NULL;
